Conflicts:
	src/main/java/Endpoints/SparkEndPoint.java

+++ src/test/java/Services/GetAllGamesFromDateServiceIT.java
@@ -52,7 +52,7 @@ public class GetAllGamesFromDateServiceIT {
      */
     @Test
     public void testExecute() {
-        Long date = 12345678L;
+        int date = 12345678;
         GameDao gd = GameDao.findById(SetUpTestObjects.getGameId1());
         Game game = new Game(gd);
         game.setDate(date);
